fix(B-464): split session turn schema and model
